+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'product/bootstrap/app_bootstrap.dart';
 
-Future<void> main() async {
-  await AppBootstrap.run();
+Future<void> main(List<String> args) async {
+  await AppBootstrap.run(args);
 }